requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-lichess==0.10
pytz==2021.3
requests==2.27.1
six==1.16.0
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
